refactor: move calendar cells logic to Cells component
