+++ src/components/Cells/Cells.styled.ts
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+export const CellsWrapper = styled.div`
+  display: flex;
+  justify-content: space-between;
+  flex-wrap: wrap;
+`;
+
+export const Cell = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 32px;
+  height: 32px;
+  font-weight: 600;
+`;
