+++ src/components/Cells/WeekCells/WeekCells.styled.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+import { Cell } from '@/components/Cells/Cells.styled';
+
+export const WeekCell = styled(Cell)`
+  font-weight: 700;
+`;
